test(logging): allow more than one module to capture logs
The directory-logging privacy test installs its capturing subscriber with
`set_global_default`, which can only ever happen once, so a second module
asserting on what it logs would panic. Move the capture into `test_support`
and install it per thread again, which is what more than one caller needs.

Doing that alone would bring back the flake `set_global_default` was working
around: `tracing` caches a callsite's interest the first time it is reached,
and a callsite first reached while the current thread has no subscriber is
cached as uninteresting for every thread, so an unrelated test enumerating a
directory could silence the DEBUG event inside a capture running beside it.
Keeping a permissive discarding subscriber installed globally for the test
binary means no thread is ever without one, so callsites stay interesting and
each event is resolved against whichever subscriber its own thread has.
